requirements.txt
pandas==1.5.3
numpy==1.24.4
requests==2.31.0
google-cloud-bigquery
google-auth
pandas-gbq
google-cloud-tasks==2.19.3